+++ tests/testthat/test-editClusterST.R
@@ -114,6 +114,21 @@ test_that("edit st-storage clusters (only for study >= v8.6.0" , {
                                                     "reservoircapacity")])
   testthat::expect_equal(list_params, value_to_test)
   
+  ### bad edit values. v8.8.0 ----
+  
+  opts_test$antaresVersion <- 880
+  wrong_list_params <- list_params
+  wrong_list_params$initiallevel <- 0.8
+  wrong_list_params$initialleveloptim <- FALSE
+  
+  expect_warning(  editClusterST(area = area_test, 
+                                 cluster_name = name_cluster_test, 
+                                 storage_parameters = wrong_list_params,
+                                 opts = opts_test, 
+                                 add_prefix = FALSE), 
+                   regexp = "`initiallevel` value will be replaced by 0.5")
+  opts_test$antaresVersion <- 860
+  
  
   ## edit DATA ----
   val <- 0.007